+++ SplunkNavigation/Sources/SplunkNavigation/Navigation+Span.swift
@@ -62,6 +62,9 @@ extension Navigation {
         if let lastScreenName = navigation.lastScreenName {
             navigationSpan.clearAndSetAttribute(key: Self.lastScreenNameKey, value: lastScreenName)
         }
+        else {
+            navigationSpan.setAttribute(key: Self.lastScreenNameKey, value: nil as AttributeValue?)
+        }
         navigationSpan.clearAndSetAttribute(key: Self.screenNameKey, value: screenName)
 
         let navigationEnd = navigation.end ?? Date()